requirements.txt

fastapi
uvicorn
qdrant-client
sentence-transformers
python-docx
pymupdf
openpyxl
xlrd
langchain
openai
langchain-community
python-dotenv
pandas
